Refactor custom random walk tutorial.
